+++ src/ui_bevy/ui_ext/text.rs
@@ -267,7 +267,7 @@ impl StaticAttribute for TextLineColor
 }
 
 impl ResponsiveAttribute for TextLineColor {}
-impl AnimatableAttribute for TextLineColor {}
+impl AnimatedAttribute for TextLineColor {}
 
 //-------------------------------------------------------------------------------------------------------------------